requirements.txt
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.8.0
lxml>=5.0.0
email-validator>=2.0.0

